day18: Workaround race condition for realz
Even after c71c062 the tests were intermittently (1/50) failing with
program1 sending the same amount as program0. The extra messages all
contained `0` values.

Prevent this by never sending `0` values. I say workaround, rather than
fix, because this isn't to spec. But it does seem to satisfy the
original behaviour and I don't want to spend any more time on it.

In doing so I also had to fix a bug in the example test for Part2
because it uses `snd` where the second argument is a value rather than a
register. I'm testing the register first because that's the most common
case in the real input and attempting to convert the string first would
be slower.

Now that this is fixed we can safely increase the buffer size and
decrease the timeout so that the tests run faster.